Update HelloWorld.java
second attempt on code style
